macos bugs fixed

+++ src/core/container/ring_buffer/ring_buffer.hpp
@@ -104,7 +104,7 @@ private:
   u64 m_head{0};
   u64 m_tail{0};
 
-  static constexpr u64 ALIGNMENT = std::max(static_cast<u64>(16), alignof(T));
+  static constexpr u64 ALIGNMENT = std::max(static_cast<u64>(16), static_cast<u64>(alignof(T)));
 
   SV_FORCE_INLINE static u64 next_pow2(u64 v) {
     if (v < 2)
